% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomoMidSag-dataset.R
\docType{data}
\name{HomoMidSag}
\alias{HomoMidSag}
\title{HomoMidSag dataset}
\format{
A data frame with 24 rows and 184 variables
}
\usage{
data(HomoMidSag)
}
\description{
2D Cartesian coordinates of 87 landmarks quantifying the skull morphology along the midsagittal plane for 24 adult modern humans.
}
\references{
Bartsch, Silvester (2019)
The ontogeny of hominid cranial form: A geometric morphometric analysis of coordinated and compensatory processes.
Master's thesis, University of Vienna.

Mitteroecker, Philipp et al. (2020)
Morphometric variation at different spatial scales: coordination and compensation in the emergence of organismal form.
\emph{Systematic Biology}, syaa007. \url{https://doi.org/10.1093/sysbio/syaa007}

Mitteroecker, Philipp et al. (2020)
Data form: Morphometric variation at different spatial scales: coordination and compensation in the emergence of organismal form.
\emph{Dryad Digital Repository}. \url{https://doi.org/10.5061/dryad.j6q573n8s}
}
